Generalise the structural solver; add a real-time WebAssembly viewer
Two fair criticisms: the solver was specific to the examples rather than
using them as test cases, and snapshots are not a viewer.

GENERALISATION

The weather was in the solver — an Insult enum with arms for snow, wind,
lightning and fire — so every new load case meant a new arm inside the
physics, and the four cases implemented were the four that existed. The
solver now knows only mechanisms: a body-force field, drag in a moving
fluid, mass accreting on upward-facing surfaces, energy conducted along a
path, a thermal field, a point impulse. Snow, wind, lightning and fire
are constructors over them. The same accretion law covers snow, rime and
volcanic ash; the same drag law covers air and water. Materials went the
same way — a closed enum of four became a struct of numbers, and a
user-defined material needs no change to the solver.

The O(n) tree pass silently assumed away redundancy, which is the general
case. Determinacy is now a property of the structure: a forest load path
is exactly solvable by statics, anything braced is not, and the redundant
case gets a matrix-free conjugate-gradient solve on the spring network.
The fast path is not an approximation of the general one — it is the
general one with an empty tie list, and the two are asserted to agree to
1e-6 on determinate structures. The redundant path is validated against
the analytic three-bar truss: 585.8 N in the middle bar against an
analytic 585.8 N, in three iterations.

Topology::from_parts lets any structure be described directly, so the
generated ones are one source of geometry rather than the only one.

VIEWER

viewer/index.html is the engine compiled to WebAssembly and driven live —
not a port, not playback. It steps in 6.5 ms, so there is ample headroom
at 60 fps. Grow a structure over decades, then load it while it stands;
members are coloured by utilisation, and traces show mass and stress over
time so a limb coming down is a step in both. Verified end to end in
headless Chromium.

Bugs found, several by the tests contradicting an expectation:

- Bracing does NOT lower peak utilisation, and does not relieve any
  particular column: load taken off the columns goes into the braces, and
  a lateral load transfers force between windward and leeward. Only the
  total falls. The solver was right; the assertion was wrong.
- Unconverged conjugate gradient was being used as if it were a solution,
  giving tie forces off by ten orders of magnitude and stresses of 1e16.
  It now falls back to the determinate answer.
- Parts touched by no spring have an empty stiffness row but a non-zero
  load — an inconsistent system CG can never satisfy. Loose litter beside
  a structure is exactly such a part, so this was the common case.
- Brace areas were absolute, so their stiffness was out of all proportion
  to members whose size depends on how much the structure has grown.
- render_tower and render_wall emitted zero-length stub members, which was
  adequate while they were only drawn and nonsense once they had to carry
  load.
- A collapsed structure produced 1e16 stresses rather than reporting
  collapse, because the density correction divides by a vanishing volume.
- usize is 32-bit on wasm32, so the 4 GB frame memory cap overflowed at
  compile time.

75 tests.

+++ src/budget.rs
@@ -106,7 +106,11 @@ pub struct FrameBudget {
     /// input and slack.
     pub sim_fraction: f64,
     /// Hard cap on materialised detail, bytes.
-    pub memory_cap: usize,
+    ///
+    /// `u64` rather than `usize` because the cap describes a *device* budget,
+    /// not the host's address space: a 32-bit target (wasm32) cannot even name
+    /// four gigabytes, and the constant overflowed at compile time there.
+    pub memory_cap: u64,
     /// Multiplicative correction learned from measured frames.
     calibration: f64,
     /// Exponential moving average of the achieved frame time.
